+++ src/server.rs
@@ -34,7 +34,7 @@ impl LanguageServer for Backend {
             capabilities: ServerCapabilities {
                 signature_help_provider: Some(SignatureHelpOptions {
                     trigger_characters: Some(vec!["(".to_string(), ",".to_string()]),
-                    retrigger_characters: Some(vec![",".to_string()]),
+                    retrigger_characters: Some(vec![",".to_string(), ")".to_string()]),
                     work_done_progress_options: WorkDoneProgressOptions {
                         work_done_progress: None,
                     },